XTA: Split CG construction and type flow into distinct sub-analyses.
