+++ OPAL/tac/src/main/scala/org/opalj/tac/fpcf/analyses/cg/xta/XTAState.scala
@@ -6,18 +6,15 @@ package analyses
 package cg
 package xta
 
-import org.opalj.br.ArrayType
 import org.opalj.br.DefinedMethod
-import org.opalj.br.Field
 import org.opalj.br.Method
 import org.opalj.br.ObjectType
 import org.opalj.br.ReferenceType
-import org.opalj.br.fpcf.properties.cg.Callees
 import org.opalj.br.fpcf.properties.cg.InstantiatedTypes
 import org.opalj.collection.immutable.UIDSet
 import org.opalj.fpcf.EOptionP
-import org.opalj.fpcf.SomeEOptionP
-import org.opalj.fpcf.SomePartialResult
+import org.opalj.fpcf.Entity
+import org.opalj.fpcf.Property
 import org.opalj.tac.fpcf.properties.TACAI
 
 import scala.collection.mutable
@@ -27,43 +24,12 @@ import scala.collection.mutable
  *
  * @author Andreas Bauer
  */
+// TODO AB code duplication: this is very similar to the RTAState (except for the entity of the type property).
 class XTAState(
-        override val method:                       DefinedMethod,
-        override protected[this] var _tacDependee: EOptionP[Method, TACAI],
-
-        private[this] var _ownInstantiatedTypesDependee: EOptionP[DefinedMethod, InstantiatedTypes],
-
-        private[this] var _calleeDependee: EOptionP[DefinedMethod, Callees],
-
-        // Field stuff
-        // TODO AB optimize...
-        private[this] var _writtenFields: Set[Field],
-        // we only need type updates of fields the method reads
-        private[this] var _readFieldTypeDependees: mutable.Map[Field, EOptionP[Field, InstantiatedTypes]],
-
-        // Array stuff
-        val methodWritesArrays: Boolean,
-        val methodReadsArrays:  Boolean
+        override val method:                          DefinedMethod,
+        override protected[this] var _tacDependee:    EOptionP[Method, TACAI],
+        private[this] var _instantiatedTypesDependee: EOptionP[DefinedMethod, InstantiatedTypes]
 ) extends CGState {
-
-    // TODO AB is there a more efficient data type for this?
-    // TODO AB maybe we also need to store the PC of the callsite here, in order to optimize for returned values which are ignored
-    private[this] var _seenCallees: mutable.Set[DefinedMethod] = mutable.Set.empty
-
-    // TODO AB is there a better data type for this?
-    private[this] var _calleeSeenTypes: mutable.LongMap[Int] = mutable.LongMap.empty
-
-    // TODO AB Can these dependees become final? Probably not!
-    // TODO AB Does this have to be a map?
-    private[this] var _calleeInstantiatedTypesDependees: mutable.Map[DefinedMethod, EOptionP[DefinedMethod, InstantiatedTypes]] = mutable.Map.empty
-
-    // NOTE AB functionally the same as calleeSeenTypes, but Field does not have an ID we can use
-    private[this] var _readFieldSeenTypes: mutable.Map[Field, Int] = mutable.Map.empty
-
-    // Note: ArrayType uses a cache internally, so identical array types will be represented by the same object.
-    private[this] var _readArraysTypeDependees: mutable.Map[ArrayType, EOptionP[ArrayType, InstantiatedTypes]] = mutable.Map.empty
-    private[this] var _readArraysSeenTypes: mutable.Map[ArrayType, Int] = mutable.Map.empty
-
     private[this] val _virtualCallSites: mutable.LongMap[mutable.Set[CallSiteT]] = mutable.LongMap.empty
 
     /////////////////////////////////////////////
@@ -72,136 +38,32 @@ class XTAState(
     //                                         //
     /////////////////////////////////////////////
 
-    def updateOwnInstantiatedTypesDependee(
-        ownInstantiatedTypesDependee: EOptionP[DefinedMethod, InstantiatedTypes]
+    def updateInstantiatedTypesDependee(
+        instantiatedTypesDependee: EOptionP[DefinedMethod, InstantiatedTypes]
     ): Unit = {
-        _ownInstantiatedTypesDependee = ownInstantiatedTypesDependee
+        _instantiatedTypesDependee = instantiatedTypesDependee
     }
 
-    def ownInstantiatedTypesDependee: Option[EOptionP[DefinedMethod, InstantiatedTypes]] = {
-        if (_ownInstantiatedTypesDependee.isRefinable)
-            Some(_ownInstantiatedTypesDependee)
+    def instantiatedTypesDependee(): Option[EOptionP[DefinedMethod, InstantiatedTypes]] = {
+        if (_instantiatedTypesDependee.isRefinable)
+            Some(_instantiatedTypesDependee)
         else
             None
     }
 
-    def ownInstantiatedTypesUB: UIDSet[ReferenceType] = {
-        if (_ownInstantiatedTypesDependee.hasUBP)
-            _ownInstantiatedTypesDependee.ub.types
+    def instantiatedTypesUB: UIDSet[ReferenceType] = {
+        if (_instantiatedTypesDependee.hasUBP)
+            _instantiatedTypesDependee.ub.types
         else
             UIDSet.empty
     }
 
     def newInstantiatedTypes(seenTypes: Int): TraversableOnce[ReferenceType] = {
-        if (_ownInstantiatedTypesDependee.hasUBP) {
-            _ownInstantiatedTypesDependee.ub.dropOldest(seenTypes)
+        if (_instantiatedTypesDependee.hasUBP) {
+            _instantiatedTypesDependee.ub.dropOldest(seenTypes)
         } else {
             UIDSet.empty
         }
-    }
-
-    // TODO AB remove once there is a better solution (e.g., split analyses)
-    private[this] var _initialPartialResults: TraversableOnce[SomePartialResult] = Seq.empty
-
-    def setInitialPartialResults(initialPartialResults: TraversableOnce[SomePartialResult]): Unit = {
-        _initialPartialResults = initialPartialResults
-    }
-
-    def initialPartialResults: TraversableOnce[SomePartialResult] = {
-        assert(_initialPartialResults.nonEmpty)
-        val results = _initialPartialResults
-        _initialPartialResults = Seq.empty
-        results
-    }
-
-    /////////////////////////////////////////////
-    //                                         //
-    //                callees                  //
-    //                                         //
-    /////////////////////////////////////////////
-
-    def calleeDependee: Option[EOptionP[DefinedMethod, Callees]] = {
-        if (_calleeDependee.isRefinable) {
-            Some(_calleeDependee)
-        } else {
-            None
-        }
-    }
-
-    def updateCalleeDependee(calleeDependee: EOptionP[DefinedMethod, Callees]): Unit = {
-        _calleeDependee = calleeDependee
-    }
-
-    def seenCallees: Set[DefinedMethod] = {
-        // TODO AB probably not very efficient
-        _seenCallees.toSet
-    }
-
-    def updateSeenCallees(newCallees: Set[DefinedMethod]): Unit = {
-        _seenCallees ++= newCallees
-    }
-
-    def calleeSeenTypes(callee: DefinedMethod): Int = {
-        _calleeSeenTypes(callee.id.toLong)
-    }
-
-    def updateCalleeSeenTypes(callee: DefinedMethod, numberOfTypes: Int): Unit = {
-        assert(numberOfTypes >= _calleeSeenTypes.getOrElse(callee.id.toLong, 0))
-        _calleeSeenTypes.update(callee.id.toLong, numberOfTypes)
-    }
-
-    def updateCalleeInstantiatedTypesDependee(
-        eps: EOptionP[DefinedMethod, InstantiatedTypes]
-    ): Unit = {
-        _calleeInstantiatedTypesDependees.update(eps.e, eps)
-    }
-
-    /////////////////////////////////////////////
-    //                                         //
-    //                 fields                  //
-    //                                         //
-    /////////////////////////////////////////////
-
-    def writtenFields: Set[Field] = _writtenFields
-
-    def updateAccessedFieldInstantiatedTypesDependee(
-        eps: EOptionP[Field, InstantiatedTypes]
-    ): Unit = {
-        _readFieldTypeDependees.update(eps.e, eps)
-    }
-
-    def fieldSeenTypes(field: Field): Int = {
-        _readFieldSeenTypes.getOrElse(field, 0)
-    }
-
-    def updateReadFieldSeenTypes(field: Field, numberOfTypes: Int): Unit = {
-        assert(numberOfTypes >= _readFieldSeenTypes.getOrElse(field, 0))
-        _readFieldSeenTypes.update(field, numberOfTypes)
-    }
-
-    /////////////////////////////////////////////
-    //                                         //
-    //                 arrays                  //
-    //                                         //
-    /////////////////////////////////////////////
-
-    def availableArrayTypes: UIDSet[ArrayType] = {
-        ownInstantiatedTypesUB collect { case at: ArrayType ⇒ at }
-    }
-
-    def updateReadArrayInstantiatedTypesDependee(
-        eps: EOptionP[ArrayType, InstantiatedTypes]
-    ): Unit = {
-        _readArraysTypeDependees.update(eps.e, eps)
-    }
-
-    def arrayTypeSeenTypes(arrayType: ArrayType): Int = {
-        _readArraysSeenTypes.getOrElse(arrayType, 0)
-    }
-
-    def updateArrayTypeSeenTypes(arrayType: ArrayType, numberOfTypes: Int): Unit = {
-        assert(numberOfTypes >= arrayTypeSeenTypes(arrayType))
-        _readArraysSeenTypes.update(arrayType, numberOfTypes)
     }
 
     /////////////////////////////////////////////
@@ -236,32 +98,13 @@ class XTAState(
     /////////////////////////////////////////////
 
     override def hasOpenDependencies: Boolean = {
-        super.hasOpenDependencies ||
-            _ownInstantiatedTypesDependee.isRefinable ||
-            _calleeDependee.isRefinable ||
-            _calleeInstantiatedTypesDependees.nonEmpty ||
-            _readFieldTypeDependees.nonEmpty ||
-            _readArraysSeenTypes.nonEmpty
+        _instantiatedTypesDependee.isRefinable || super.hasOpenDependencies
     }
 
-    override def dependees: List[SomeEOptionP] = {
-        var dependees = super.dependees
-
-        if (ownInstantiatedTypesDependee.isDefined)
-            dependees ::= ownInstantiatedTypesDependee.get
-
-        if (calleeDependee.isDefined)
-            dependees ::= calleeDependee.get
-
-        if (_calleeInstantiatedTypesDependees.nonEmpty)
-            dependees ++= _calleeInstantiatedTypesDependees.values
-
-        if (_readFieldTypeDependees.nonEmpty)
-            dependees ++= _readFieldTypeDependees.values
-
-        if (_readArraysTypeDependees.nonEmpty)
-            dependees ++= _readArraysTypeDependees.values
-
-        dependees
+    override def dependees: List[EOptionP[Entity, Property]] = {
+        if (instantiatedTypesDependee().isDefined)
+            instantiatedTypesDependee().get :: super.dependees
+        else
+            super.dependees
     }
 }
